+++ database/migrations/2026_06_27_165025_create_portfolio_items_table.php
@@ -11,25 +11,27 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('portfolio_items', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->foreignId('portfolio_category_id')->constrained()->onDelete('cascade');
-            $table->foreignId('client_id')->nullable()->constrained()->onDelete('set null');
-            $table->foreignId('project_id')->nullable()->constrained()->onDelete('set null');
-            $table->string('title');
-            $table->string('slug');
-            $table->text('description');
-            $table->string('technologies')->nullable();
-            $table->string('redirect_url')->nullable();
-            $table->string('thumb_path')->nullable();
-            $table->string('status')->default('rascunho'); // rascunho, publicado
-            $table->boolean('is_featured')->default(false);
-            $table->string('meta_title')->nullable();
-            $table->text('meta_description')->nullable();
-            $table->string('meta_keywords')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('portfolio_items')) {
+            Schema::create('portfolio_items', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->foreignId('portfolio_category_id')->constrained()->onDelete('cascade');
+                $table->foreignId('client_id')->nullable()->constrained()->onDelete('set null');
+                $table->foreignId('project_id')->nullable()->constrained()->onDelete('set null');
+                $table->string('title');
+                $table->string('slug');
+                $table->text('description');
+                $table->string('technologies')->nullable();
+                $table->string('redirect_url')->nullable();
+                $table->string('thumb_path')->nullable();
+                $table->string('status')->default('rascunho'); // rascunho, publicado
+                $table->boolean('is_featured')->default(false);
+                $table->string('meta_title')->nullable();
+                $table->text('meta_description')->nullable();
+                $table->string('meta_keywords')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
